refactor: split marketplace commands into package modules
